insert cart to database

+++ lib/models/book.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 // The model of a book , parameters a book must contain
 class Book {
-  final double bookId;
+  final int bookId;
   final String title;
   final String author;
   final double rating;
